Add `EmptyFragment`
