+++ app/src/main/res/values/styles.xml
@@ -7,21 +7,13 @@
 
     <style name="text_view_style.header">
         <item name="android:textColor">@color/black</item>
-        <item name="android:textSize">@dimen/text_size_medium</item>
+        <item name="android:textSize">@dimen/text_size_default</item>
         <item name="android:textStyle">bold</item>
     </style>
 
     <style name="text_view_style.center">
         <item name="android:layout_width">match_parent</item>
         <item name="android:gravity">center_horizontal</item>
-    </style>
-
-    <style name="text_view_style.outline">
-        <item name="android:shadowColor">#000000</item>
-        <item name="android:shadowDx">1.7</item>
-        <item name="android:shadowDy">1.5</item>
-        <item name="android:shadowRadius">1.8</item>
-        <item name="android:textColor">@color/white</item>
     </style>
 
     <style name="card_view_style">
@@ -44,7 +36,17 @@
         <item name="android:layout_marginRight">@dimen/size_default</item>
     </style>
 
-    <style name="ShapeAppearanceOverlay.Rounded">
+    <style name="ShapeAppearanceOverlay.App.CornerSize50Percent" parent="">
         <item name="cornerSize">50%</item>
+    </style>
+
+    <style name="text_view_style.toolbar_expanded">
+        <item name="android:textSize">@dimen/text_size_big</item>
+        <item name="android:shadowColor">#000000</item>
+        <item name="android:shadowDx">1.7</item>
+        <item name="android:shadowDy">1.5</item>
+        <item name="android:shadowRadius">1.8</item>
+        <item name="android:textAllCaps">true</item>
+        <item name="android:textColor">@color/white</item>
     </style>
 </resources>
